Split growing supplied product builder
